+++ assembly/onpremises-ide-packaging-war-platform-api/src/main/java/com/codenvy/api/account/personal/PersonalAccountModule.java
@@ -1,0 +1,40 @@
+/*
+ *  [2012] - [2017] Codenvy, S.A.
+ *  All Rights Reserved.
+ *
+ * NOTICE:  All information contained herein is, and remains
+ * the property of Codenvy S.A. and its suppliers,
+ * if any.  The intellectual and technical concepts contained
+ * herein are proprietary to Codenvy S.A.
+ * and its suppliers and may be covered by U.S. and Foreign Patents,
+ * patents in process, and are protected by trade secret or copyright law.
+ * Dissemination of this information or reproduction of this material
+ * is strictly forbidden unless prior written permission is obtained
+ * from Codenvy S.A..
+ */
+package com.codenvy.api.account.personal;
+
+import com.codenvy.resource.api.free.DefaultResourcesProvider;
+import com.codenvy.resource.api.usage.ResourcesPermissionsChecker;
+import com.google.inject.AbstractModule;
+import com.google.inject.multibindings.Multibinder;
+
+import org.eclipse.che.api.user.server.UserManager;
+import org.eclipse.che.inject.DynaModule;
+
+/**
+ * @author Sergii Leschenko
+ */
+@DynaModule
+public class PersonalAccountModule extends AbstractModule {
+    @Override
+    protected void configure() {
+        Multibinder.newSetBinder(binder(), DefaultResourcesProvider.class)
+                   .addBinding().to(DefaultUserResourcesProvider.class);
+
+        Multibinder.newSetBinder(binder(), ResourcesPermissionsChecker.class)
+                   .addBinding().to(UserResourcesPermissionsChecker.class);
+
+        bind(UserManager.class).to(OnpremisesUserManager.class);
+    }
+}
