+++ assembly/onpremises-ide-packaging-war-platform-api/src/main/java/com/codenvy/api/deploy/OnPremisesIdeApiModule.java
@@ -49,8 +49,6 @@ import com.codenvy.plugin.github.factory.resolver.GithubFactoryParametersResolve
 import com.codenvy.plugin.gitlab.factory.resolver.GitlabFactoryParametersResolver;
 import com.codenvy.report.ReportModule;
 import com.codenvy.resource.api.ResourceModule;
-import com.codenvy.resource.api.free.DefaultResourcesProvider;
-import com.codenvy.resource.api.usage.ResourcesPermissionsChecker;
 import com.codenvy.service.bitbucket.BitbucketConfigurationService;
 import com.codenvy.service.systemram.DockerBasedSystemRamInfoProvider;
 import com.codenvy.service.systemram.SystemRamInfoProvider;
@@ -100,7 +98,6 @@ import org.eclipse.che.api.ssh.server.jpa.SshJpaModule;
 import org.eclipse.che.api.user.server.PreferencesService;
 import org.eclipse.che.api.user.server.ProfileService;
 import org.eclipse.che.api.user.server.TokenValidator;
-import org.eclipse.che.api.user.server.UserManager;
 import org.eclipse.che.api.user.server.UserService;
 import org.eclipse.che.api.user.server.jpa.UserJpaModule;
 import org.eclipse.che.api.workspace.server.WorkspaceConfigMessageBodyAdapter;
@@ -150,14 +147,6 @@ public class OnPremisesIdeApiModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        Multibinder.newSetBinder(binder(), DefaultResourcesProvider.class)
-                   .addBinding().to(DefaultUserResourcesProvider.class);
-
-        Multibinder.newSetBinder(binder(), ResourcesPermissionsChecker.class)
-                   .addBinding().to(UserResourcesPermissionsChecker.class);
-
-        bind(UserManager.class).to(OnpremisesUserManager.class);
-
         bind(ApiInfoService.class);
         bind(ProjectTemplateRegistry.class);
         bind(ProjectTemplateDescriptionLoader.class).asEagerSingleton();
